pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai>=0.134.0",
    "crewai-tools>=0.48.0",
    "fastapi>=0.115.14",
    "google-generativeai>=0.8.5",
    "langchain-community>=0.3.26",
    "pydantic>=2.11.7",
    "pypdf>=5.7.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "uvicorn[standard]>=0.35.0",
]
